requirements.txt
altgraph==0.17.4
bottle==0.13.2
bottle-websocket==0.2.9
cffi==1.17.1
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Eel==0.17.0
et-xmlfile==1.1.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.2.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
humanfriendly==10.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
onnx==1.13.1
onnxruntime==1.15.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pefile==2023.2.7
pillow==10.2.0
protobuf==3.20.3
pycocotools==2.0.7
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.1.4
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
six==1.16.0
sympy==1.13.1
typing_extensions==4.9.0
whichcraft==0.6.1
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
